+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## _(in-progress)_
+* Change: Note compatibility through WP 5.3+
+
 ## 2.5 _(2019-04-04)_
 
 ### Highlights: