+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Change: Add inline documentation for class variables
 * Change: Add link to CHANGELOG.md in README.md
 * Change: Update copyright date (2020)
+* Delete: Remove unused class variable `$config`
 
 ## 2.5 _(2019-04-04)_